Added validator tests on rax:code extension.
